+++ controller/book_controller.go
@@ -1,4 +1,4 @@
-package service
+package controller
 
 import (
 	"github.com/gin-gonic/gin"
