+++ controller/error.go
@@ -1,4 +1,4 @@
-package service
+package controller
 
 import (
 	"errors"
